NEW Interfaces in DataObjectScaffolder (fixes #209)
This enables more stable query creation with union types.
Instead of referring to specific types (e.g. "... on Page"),
you can use the relevant interface for that type (e.g. "...on PageInterface").
Since all types in the union share those interfaces, this results in less duplication
of fields in the query. It also makes the query resilient against adding new types
to the union (or subclasses to the DataObject), which would previously result in NULL query results.